% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{activateImage,MeaData-method}
\alias{activateImage,MeaData-method}
\alias{activeImage,MeaData-method}
\title{Default image name}
\usage{
\S4method{activateImage}{MeaData}(object, img_name, unload = TRUE)

\S4method{activeImage}{MeaData}(object)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image to activate as
the default image.}
}
\value{
\itemize{
\item{activateImage()}: Updated \code{SPATA2} object.
\item{activeImage()}: Character value. Name of the currently active image. Empty string if no images exist.
}
}
\description{
Sets and extracts the active (default) image name.
}
