% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{activeMea}
\alias{activeMea}
\alias{activateMea}
\title{Default MEA name}
\usage{
activeMea(object)

activateMea(object, mea_name, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\itemize{
\item{activateMea()}: Updated \code{SPATA2} object.
\item{activeMea()}: Character value. Name of the currently MEA data. Empty string if no MEA data exists.
}
}
\description{
Sets and extracts the active (default) \link[=MeaData]{MEA} name.
}
\seealso{
\code{\link[=getMeaData]{getMeaData()}}, \code{\link[=getMeaNodes]{getMeaNodes()}}, \code{\link[=getMeaEdges]{getMeaEdges()}}
}
