% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{addVarsToEdges}
\alias{addVarsToEdges}
\alias{addVarsToEdges,SPATA2-method}
\alias{addVarsToEdges,MeaData-method}
\alias{addVarsToNodes}
\alias{addVarsToNodes,SPATA2-method}
\alias{addVarsToNodes,MeaData-method}
\title{Add variables to MEA edges or nodes}
\usage{
addVarsToEdges(object, ...)

\S4method{addVarsToEdges}{SPATA2}(
  object,
  edges_data,
  var_names = NULL,
  mea_name = activeMea(object),
  overwrite = FALSE
)

\S4method{addVarsToEdges}{MeaData}(object, edges_data, var_names = NULL, overwrite = FALSE)

addVarsToNodes(object, ...)

\S4method{addVarsToNodes}{SPATA2}(
  object,
  node_data,
  var_names = NULL,
  mea_name = activeMea(object),
  overwrite = FALSE
)

\S4method{addVarsToNodes}{MeaData}(object, node_data, var_names = NULL, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}, an object that contains MEA data
or, in case of S4 generics, objects of classes for which a methods has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{edges_data}{A data.frame containing new or updated variables for MEA edges. The \code{id} column is required
to match rows with the existing edge data.}

\item{var_names}{\emph{character} or \code{NULL}. The names of the variables to be added or updated. If \code{NULL} (default),
all columns in \code{edges_data} or \code{node_data} (excluding \code{id}) are used.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{node_data}{A data.frame containing new or updated variables for MEA nodes. The \code{id} column is required
to match rows with the existing node data.}
}
\description{
These functions add new variables or update existing ones in the MEA edge or node data
of the provided object. The variables are added by merging the existing edge or node data with
the supplied data frames (\code{edges_data} or \code{node_data}).
}
\details{
\itemize{
\item \strong{\code{addVarsToEdges}}: Adds or updates variables in the MEA edge data.
\item \strong{\code{addVarsToNodes}}: Adds or updates variables in the MEA node data.
\item For both functions, the \code{id} column is used to align the input data with the existing edge or node data.
}
}
\seealso{
\code{\link[=getMeaEdges]{getMeaEdges()}}, \code{\link[=getMeaNoes]{getMeaNoes()}}, \code{\link[=setMeaEdges]{setMeaEdges()}}, \code{\link[=setMeaNodes]{setMeaNodes()}}
}
