% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{adjust_ids}
\alias{adjust_ids}
\title{Adjust MEA IDs}
\usage{
adjust_ids(df, mea_design, verbose = FALSE)
}
\arguments{
\item{df}{Nodes data.frame of the respective mea design.}

\item{mea_design}{Character value. The name of the MEA design/platform.}
}
\description{
Ensures valid electrode/node IDS.
}
\keyword{internal}
