% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{computeDistToNME}
\alias{computeDistToNME}
\title{Compute distance to nearest MEA electrode (NME)}
\usage{
computeDistToNME(
  object,
  nme_name,
  nme_dist_name,
  mea_name = activeMea(object),
  unit = getDefaultUnit(object),
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{nme_name}{Character value. The name of the metadata variable indicating the nearest
MEA node for each observation - as identified by \code{\link[=identifyNME]{identifyNME()}}.}

\item{nme_dist_name}{Character value. The name of the metadata variable to store the computed distances.}

\item{unit}{The unit for the computed distances (e.g., \code{"um"}, \code{"mm"}).
Defaults to the object's default unit (\code{getDefaultUnit(object)}).}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\description{
This function computes the distance between observations in the \code{\link{SPATA2}} object
and their \link[=identifyNME]{nearest MEA electrode}. The computed distances
are added as a new metadata variable in the object.
}
\seealso{
\code{\link[=addMetaDataObs]{addMetaDataObs()}}, \code{\link[=getMeaNodes]{getMeaNodes()}}, \code{\link[=getCoordsDf]{getCoordsDf()}}, \code{\link[=getCoordsMtr]{getCoordsMtr()}}
}
