% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{containsHistoImages,MeaData-method}
\alias{containsHistoImages,MeaData-method}
\title{Check availability of image containers}
\usage{
\S4method{containsHistoImages}{MeaData}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the input object contains any \code{\link[SPATA2]{HistoImage}} objects.

Note to confuse with \code{\link[SPATA2:containsImage]{containsImage()}}.
}
