% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{containsMEA}
\alias{containsMEA}
\title{Check availability of an image}
\usage{
containsMEA(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the input object has one or more objects of class \code{\link{MeaData}} stored
in the respective slot.
}
