% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{createMeaObjectEmpty}
\alias{createMeaObjectEmpty}
\title{Create an empty MEA data object}
\usage{
createMeaObjectEmpty(name, mea_design, ..., verbose = TRUE)
}
\arguments{
\item{name}{Character value. Name of the output object (as set in slot \verb{@name}).}

\item{mea_design}{Character value or \code{\link{MeaDesign}}. The design of the MEA platform.
If character, must be one of \code{validMeaDesigns()}.}
}
\value{
An object of class \code{\link{MeaData}}.
}
\description{
Creates an empty object of class \code{\link{MeaData}}.
}
