% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{eDiam}
\alias{eDiam}
\alias{eShape}
\alias{eSpacing}
\alias{eSpacing,SPATA2-method}
\alias{eSpacing,MeaData-method}
\title{Obtain electrode information}
\usage{
eDiam(object)

eShape(object)

eSpacing(object, ...)

\S4method{eSpacing}{SPATA2}(object, unit = getDefaultUnit(object), img_name = activeImage(object))

\S4method{eSpacing}{MeaData}(object, unit = getDefaultUnit(object, which = "dist"))
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}, an object that contains MEA data
or, in case of S4 generics, objects of classes for which a methods has been defined.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}
}
\description{
Functions to extract hardwired information about the electrodes of the
\code{\link{MeaDesign}}.
}
