% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{expand_mea_grid}
\alias{expand_mea_grid}
\title{Expand MEA Grid}
\usage{
expand_mea_grid(vertices, n_rows, n_cols)
}
\arguments{
\item{vertices}{A tibble/data.frame carrying spatial information about the vertices of the grid, with 3 columns:
\itemize{
\item \code{vertex} (character): must include the four vertex labels \emph{'bl'}, \emph{'br'}, \emph{'tl'}, and \emph{'tr'}.
\item \code{x} (numeric): x-coordinates of the vertices.
\item \code{y} (numeric): y-coordinates of the vertices.
}}

\item{n_rows}{Number of rows in the grid. Must be 2 or higher.}

\item{n_cols}{Number of columns in the grid. Must be 2 or higher.}
}
\value{
A tibble containing the interpolated grid points with columns:
\itemize{
\item \code{id}: Unique grid point identifier in the form \code{x<col>y<row>}.
\item \code{col}: Column index of the grid point.
\item \code{row}: Row index of the grid point.
\item \code{x}: x-coordinate of the grid point.
\item \code{y}: y-coordinate of the grid point.
}
}
\description{
This function interpolates a grid of points within a quadrilateral defined by four vertices.
}
\details{
The vertices of \code{vertices} are named according to their relative positions:
\itemize{
\item \strong{l}: Left
\item \strong{r}: Right
\item \strong{t}: Top
\item \strong{b}: Bottom
}
Therefore, the vertices are named:
\itemize{
\item \code{bl}: Bottom Left
\item \code{br}: Bottom Right
\item \code{tl}: Top Left
\item \code{tr}: Top Right
}
The function linearly interpolates points along the left and right edges of the quadrilateral and
then interpolates grid points between these edges.
}
\examples{
vertices <- tibble::tibble(
  vertex = c("lt", "rt", "rb", "lb"),
  x = c(598, 578, 286, 305),
  y = c(723, 1019, 997, 702)
)
expand_mea_grid(vertices, n_rows = 4, n_cols = 4)
}
