% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getHistoImage,MeaData-method}
\alias{getHistoImage,MeaData-method}
\title{Obtain HistoImage object}
\usage{
\S4method{getHistoImage}{MeaData}(object, img_name = activeImage(object), ...)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[SPATA2:concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[SPATA2:getImageNames]{getImageNames()}}.}

\item{...}{}
}
\description{
Extracts the S4-containers of registered images. Note that
slot @image might be empty. Use \code{loadImage()} in that case.

\itemize{
\item{\code{getHistoImage()}:}{ Extracts object by name. By default, the active \code{HistoImage} image is returned.}
\item{\code{getHistoImageRef()}:}{ Extracts the reference \code{HistoImage} object.}
}
}
