% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaDesign}
\alias{getMeaDesign}
\alias{getMeaDesign,SPATA2-method}
\alias{getMeaDesign,MeaData-method}
\title{Obtain MEA design}
\usage{
getMeaDesign(object, ...)

\S4method{getMeaDesign}{SPATA2}(object, mea_name = activeMea(object))

\S4method{getMeaDesign}{MeaData}(object)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}
}
\value{
An object of class \code{\link{MeaDesign}}.
}
\description{
Extracts an object of class \code{\link{MeaDesign}}.
}
