% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaEdges}
\alias{getMeaEdges}
\alias{getMeaEdges,SPATA2-method}
\alias{getMeaEdges,MeaData-method}
\title{Obtain MEA edges}
\usage{
getMeaEdges(object, ...)

\S4method{getMeaEdges}{SPATA2}(
  object,
  mea_name = activeMea(object),
  img_name = activeImage(object)
)

\S4method{getMeaEdges}{MeaData}(object, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or \code{\link{MeaData}}.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}
}
\value{
A \code{data.frame} containing edge information with the following columns:
\itemize{
\item \emph{id}: Edge ID.
\item \emph{mea_name}: Name of \code{\link{MeaData}} object.
\item \emph{sample}: Name of tissue sample.
\item \code{from}: Starting node (electrode ID).
\item \code{to}: Ending node (electrode ID).
\item \code{x}: Scaled x-coordinate of the starting node.
\item \code{y}: Scaled y-coordinate of the starting node.
\item \code{xend}: Scaled x-coordinate of the ending node.
\item \code{yend}: Scaled y-coordinate of the ending node.
\item Additional columns based on the specific MEA edge data.
}
}
\description{
Extracts the edge data for the specified Microelectrode Array (MEA). The edges
represent connections between electrodes (nodes) in the MEA and include spatial coordinates.
Coordinates are scaled based on the associated image's scaling factor.
}
\details{
\itemize{
\item \strong{SPATA2 Method}: Delegates the operation to the corresponding \code{MeaData} object within the \code{SPATA2} object.
\item \strong{MeaData Method}:
\itemize{
\item Retrieves the edge data and scales the spatial coordinates (\code{x}, \code{y}, \code{xend}, \code{yend})
based on the scaling factor of the associated image.
\item The original spatial coordinates (\code{x_orig}, \code{y_orig}, \code{xend_orig}, \code{yend_orig}) are retained.
}
}
}
\examples{
# Retrieve edges from a SPATA2 object
edges <- getMeaEdges(object)

# Retrieve edges from a MeaData object with a specific image
edges <- getMeaEdges(mea_data, img_name = "example_image")

}
\seealso{
\code{\link[=getMeaNodes]{getMeaNodes()}}, \code{\link[=getMeaData]{getMeaData()}}, \code{\link[=getImageScaleFactor]{getImageScaleFactor()}}
}
