% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaGraph}
\alias{getMeaGraph}
\alias{getMeaGraph,SPATA2-method}
\alias{getMeaGraph,MeaData-method}
\title{Obtain MEA data as a graph}
\usage{
getMeaGraph(object, ...)

\S4method{getMeaGraph}{SPATA2}(
  object,
  mea_name = activeMea(object),
  img_name = activeImage(object),
  tidy = TRUE
)

\S4method{getMeaGraph}{MeaData}(object, img_name = activeImage(object), tidy = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}, an object that contains MEA data
or, in case of S4 generics, objects of classes for which a methods has been defined.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}
}
\value{
A graph object.
}
\description{
Extracts nodes and edges of the MEA data as a graph object.
}
