% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaNames}
\alias{getMeaNames}
\title{Obtain MEA data names}
\usage{
getMeaNames(object)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts the names of all \link[=MeaData]{MEA} data objects in the object.
}
