% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaNodes}
\alias{getMeaNodes}
\alias{getMeaNodes,SPATA2-method}
\alias{getMeaNodes,MeaData-method}
\title{Obtain MEA nodes}
\usage{
getMeaNodes(object, ...)

\S4method{getMeaNodes}{SPATA2}(
  object,
  mea_name = activeMea(object),
  img_name = activeImage(object),
  complete = FALSE,
  ...
)

\S4method{getMeaNodes}{MeaData}(object, img_name = activeImage(object), complete = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or \code{\link{MeaData}}.}

\item{complete}{\emph{logical} (optional). If \code{FALSE} (default), filters out nodes marked as missing.}

\item{inherit}{mea_doc_dummy params}
}
\value{
A \code{data.frame} containing node information with the following columns:
\itemize{
\item \code{id}: Node identifier.
\item \code{mea_name}: Name of the MEA.
\item \code{sample}: Sample identifier.
\item \code{x}: X-coordinate (scaled).
\item \code{y}: Y-coordinate (scaled).
\item Additional columns based on the specific MEA data.
}
}
\description{
Extracts the node data for the specified Microelectrode Array (MEA). The nodes represent
electrodes in the MEA grid, and their spatial coordinates can be adjusted to account for the
scaling factors of associated images.
}
\examples{
# Retrieve nodes from a SPATA2 object
nodes <- getMeaNodes(object)

# Retrieve nodes from a MeaData object with specific settings
nodes <- getMeaNodes(mea_data, img_name = "example_image", complete = TRUE)

}
\seealso{
\code{\link[=getImageScaleFactor]{getImageScaleFactor()}}, \code{\link[=getMeaData]{getMeaData()}}
}
