% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getMeaOutline}
\alias{getMeaOutline}
\alias{getMeaOutline,SPATA2-method}
\alias{getMeaOutline,MeaData-method}
\title{Obtain MEA outline}
\usage{
getMeaOutline(object, ...)

\S4method{getMeaOutline}{SPATA2}(object, img_name = activeImage(object), ...)

\S4method{getMeaOutline}{MeaData}(object, img_name = activeImage(object), ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or \code{\link{MeaData}}.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}
}
\value{
A \code{data.frame} containing the outline information with the following columns:
\itemize{
\item \code{vertex}: Vertex label (e.g., \code{bl}, \code{tl}, \code{tr}, \code{br}).
\item \code{x}: X-coordinate (scaled).
\item \code{y}: Y-coordinate (scaled).
\item Additional columns based on the specific MEA design.
}
}
\description{
Extracts the outline of the specified Microelectrode Array (MEA). The outline defines
the spatial boundary of the MEA, and its coordinates can be adjusted to account for the scaling
factors of associated images.
}
\examples{
# Retrieve outline from a SPATA2 object
outline <- getMeaOutline(object)

# Retrieve outline from a MeaData object with specific settings
outline <- getMeaOutline(mea_data, img_name = "example_image")

}
\seealso{
\code{\link[=getImageScaleFactor]{getImageScaleFactor()}}, \code{\link[=getMeaData]{getMeaData()}}, \code{\link[=getMeaDesign]{getMeaDesign()}}
}
