% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{getPixelScaleFactor,MeaData-method}
\alias{getPixelScaleFactor,MeaData-method}
\title{Obtain scale factor for pixel to SI conversion}
\usage{
\S4method{getPixelScaleFactor}{MeaData}(
  object,
  img_name = activeImage(object),
  unit = getDefaultUnit(object, "dist"),
  switch = FALSE,
  add_attr = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[SPATA2:concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[SPATA2:getImageNames]{getImageNames()}}.}

\item{unit}{Character value. The SI-unit of interest.
Determines the reference unit for the pixel size.}

\item{switch}{Logical value. If \code{TRUE}, the unit of the output is switched.
See details for more.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A single numeric value with the unit defined in attribute \emph{unit}.
}
\description{
Extracts side length of pixel sides depending
on the resolution of the chosen image.
}
