% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{ggpLayerMeaEdges}
\alias{ggpLayerMeaEdges}
\title{Create ggplot2 layer for MEA edges}
\usage{
ggpLayerMeaEdges(
  object,
  edges_subset = NULL,
  alpha_by = NULL,
  color_by = NULL,
  size_by = NULL,
  edge_alpha = 0.9,
  edge_color = "black",
  edge_size = 1,
  mea_name = activeMea(object),
  img_name = activeImage(object),
  transform_with = NULL,
  clrp = "default",
  clrp_adjust = NULL,
  clrsp = "inferno",
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}} or \code{\link{SPATA2}}.}

\item{edges_subset}{Character vector or \code{NULL}. If character, specifies the
edges to be displayed by ID. If \code{NULL}, the default, all edges are displayed.}

\item{alpha_by}{Character value or \code{NULL}. Variable name for transparency mapping.}

\item{color_by}{Character value or \code{NULL}. Variable name for color mapping.}

\item{size_by}{Character value or \code{NULL}. Variable name for size (linewidth) mapping.}

\item{edge_alpha}{Numeric value. Default transparency for edges. Used when \code{alpha_by} is \code{NULL}.}

\item{edge_color}{Character value. Default color for edges. Used when \code{color_by} is \code{NULL}.}

\item{edge_size}{Numeric value. Default thickness for edges. Used when \code{size_by} is \code{NULL}.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}

\item{transform_with}{List or \code{NULL.} If list, can be used to transform continuous variables before usage.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
E.g if the variable of interest is \emph{GFAP} gene expression, the following would work:

\itemize{
\item{Single function:}{ \code{transform_with = log10}},
\item{Multiple functions:}{ \verb{transform_with = list(GFAP = list(log10, log2)}}
}

In case of plotting:
Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{...}{Additional arguments passed to \verb{scale_color_<color_input>}.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
This function generates a \code{ggplot2} layer (\code{geom_segment}) for visualizing
edges of a microelectrode array (MEA) in a 2D plot. The appearance of the edges
can be customized by mapping edge properties (e.g., alpha, color, size) to specific variables.
}
