% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{ggpLayerMeaOutline}
\alias{ggpLayerMeaOutline}
\title{Create ggplot2 layer for MEA outline}
\usage{
ggpLayerMeaOutline(
  object,
  opt = "outline",
  img_name = activeImage(object),
  line_alpha = 1,
  line_color = "black",
  line_size = 1,
  expand_outline = NULL,
  label_vertices = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or \code{\link{MeaData}}.}

\item{opt}{Character vector. Option \emph{'outline'}, outlines the MEA with a line. Option \emph{'crop'} zooms in.
Both can be specified simultaneously.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}

\item{line_alpha}{\emph{numeric} (optional). Transparency of the outline. Default is \code{1} (fully opaque).}

\item{line_color}{\emph{character} (optional). Color of the outline. Default is \code{"black"}.}

\item{line_size}{\emph{numeric} (optional). Thickness of the outline. Default is \code{1}.}

\item{expand_outline}{\emph{numeric} (optional). If provided, the outline will be expanded by the
buffer distance (in the same units as the MEA coordinates). Default is \code{NULL}.}

\item{label_vertices}{Logical value. If \code{TRUE}, the labels vertices used to outline the
electrode array are displayed. Note that \code{expand_outline} does not affect the position of
the labels but only the outline, if drawn.}

\item{...}{Additional arguments passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}} if \code{label_vertices = TRUE}.}
}
\value{
A \code{ggplot2} layer (\code{geom_polygon}) representing the MEA outline.
}
\description{
This function generates a \code{ggplot2} layer (\code{geom_polygon}) to visualize the outline
of a Microelectrode Array (MEA) in a 2D plot. The outline can optionally be expanded by a
specified buffer distance.
}
\details{
\itemize{
\item The outline is extracted using the \code{getMeaOutline()} function and represents the spatial
boundary of the MEA.
\item If \code{expand_outline} is provided, the \code{buffer_area()} function is used to enlarge the outline
by the specified distance.
\item The outline is drawn as a polygon with customizable transparency, color, and line thickness.
}
}
\seealso{
\code{\link[=getMeaOutline]{getMeaOutline()}}
}
