% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{identifyNME}
\alias{identifyNME}
\title{Identify nearest MEA electrode (NME)}
\usage{
identifyNME(
  object,
  ...,
  nme_name,
  max_dist = eSpacing(object, "px")/2,
  mea_name = activeMea(object),
  complete = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Expressions that return a logical value, used to subset the \link[concept_mea_nodes]{MEA nodes}
prior to the neighbor identification.}

\item{nme_name}{Character value. The name of the new metadata variable that will store the
nearest electrode IDs.}

\item{max_dist}{\link[=concept_distance_measure]{Distance measure}. Maximum allowable distance for assigning a
nearest electrode. Observations beyond this distance are marked as "out-of-range". Defaults to half of
the average electrode spacing.}

\item{complete}{Logical value. If \code{TRUE}, the default, electrodes with no data associated are included
as potential neighbors.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function identifies the nearest electrode for each observation
in the \code{\link{SPATA2}} object and adds the information as a metadata variable. Observations
that are farther than a specified maximum distance (\code{max_dist}) are assigned an
"out-of-range" value.
}
\details{
This function identifies the nearest electrode (NME) for each \link[=concept_observations]{observation}
in the \code{\link{SPATA2}} object. A filtering step prior to the neighbor-mapping allows subsetting of electrodes
based on logical conditions, which can exclude irrelevant electrodes from the neighbor
assignment process via \code{...}. Observations farther than \code{max_dist} from any valid neighbor are labeled with \emph{'none'}.
The choice of filtering directly affects the pool of potential neighbors,
potentially increasing the likelihood of observations being labeled "out-of-range" if too
many electrodes are excluded.

The \code{complete} argument determines whether all electrodes or only those with associated data
are considered:
\itemize{
\item \strong{\code{TRUE} (default)}: Includes all electrodes, regardless of associated data, ensuring that
spatial relationships use the entire MEA layout.
\item \strong{\code{FALSE}}: Considers only electrodes with associated data, focusing the neighbor search
on relevant subsets of electrodes.
}

See \code{getMeaNodes(..., complete = TRUE/FALSE)}.
}
\seealso{
\code{\link[=addMetaDataObs]{addMetaDataObs()}}, \code{\link[=getMeaNodes]{getMeaNodes()}}, \code{\link[=computeDistToNMN]{computeDistToNMN()}}
}
