% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{manage_aesthetics}
\alias{manage_aesthetics}
\title{Manage ggplot2 aesthetics and parameters}
\usage{
manage_aesthetics(
  alpha_by = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  alpha = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  mapping = ggplot2::aes()
)
}
\arguments{
\item{alpha_by}{Character value. Column name for transparency mapping.}

\item{color_by}{Character value. Column name for color mapping.}

\item{shape_by}{Character value. Column name for size mapping.}

\item{size_by}{Character value. Column name for size mapping.}

\item{alpha}{Numeric value. Default alpha value if \code{alpha_by} is not provided.}

\item{color}{Character value. Default color if \code{color_by} is not provided.}

\item{shape}{Character value. Default shape if \code{shape_by} is not provided.}

\item{size}{Numeric value. Default size value if \code{size_by} is not provided.}

\item{mapping}{list. Mapping to start with.}
}
\value{
A list containing:
\describe{
\item{\code{mapping}}{A ggplot2 aesthetics mapping object.}
\item{\code{params}}{A named list of unused parameters to pass directly to \verb{ggplot2::geom_*}.}
}
}
\description{
This function manages the setup of aesthetics (\code{aes}) for \code{ggplot2}
layers and returns both the \code{mapping} and unused parameters as a list.
}
