% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{mea_doc_dummy}
\alias{mea_doc_dummy}
\title{Synaptiq documentation dummy}
\usage{
mea_doc_dummy()
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}, an object that contains MEA data
or, in case of S4 generics, objects of classes for which a methods has been defined.}

\item{edges}{A data.frame in which each observation corresponds to
an edge between a graph in which nodes represent the \link[=concept_electrode]{MEA-electrodes}.}

\item{alpha_by}{Character value or \code{NULL}. Variable name for transparency mapping.}

\item{color_by}{Character value or \code{NULL}. Variable name for color mapping.}

\item{edge_alpha}{Numeric value. Default transparency for edges. Used when \code{alpha_by} is \code{NULL}.}

\item{edge_color}{Character value. Default color for edges. Used when \code{color_by} is \code{NULL}.}

\item{edge_size}{Numeric value. Default thickness for edges. Used when \code{size_by} is \code{NULL}.}

\item{img_name}{Character value. The name of the image of interest - to which spatial
coordinates are scaled. Must be one of \code{\link[=getImageNames]{getImageNames()}}. By default, the active image is chosen.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{node_alpha}{Numeric value. Default transparency for nodes Used when \code{alpha_by} is \code{NULL}.}

\item{node_color}{Character value. Default color for nodes. Used when \code{color_by} is \code{NULL}.}

\item{node_shape}{Numeric value. Shape of nodes.}

\item{node_size}{Numeric value. Size of nodes.}

\item{nodes}{A data.frame in which each observation corresponds to
an \link[=concept_electrode]{electrode} in a MEA grid - conceptualized as
nodes in a graph.}

\item{size_by}{Character value or \code{NULL}. Variable name for size (linewidth) mapping.}
}
\description{
Synaptiq documentation dummy
}
