% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\docType{class}
\name{MeaData-class}
\alias{MeaData-class}
\alias{MeaData}
\alias{show,MeaData-method}
\title{Class \code{MeaData}}
\usage{
\S4method{show}{MeaData}(object)
}
\description{
A class for storing MEA experimental data, including the \link[=MeaDesign]{MEA hardware design}, electrode-level
data, stimulation events, and metadata. In downstream analysis electrode data is conceptualized as a graph
network where \emph{nodes} represent electrodes and \emph{edges} represent their connections.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(MeaData)}: Show method for the MeaData class

}}
\section{Slots}{

\describe{
\item{\code{design}}{\code{MeaDesign}. The MEA design setup as a \code{\link{MeaDesign}} object.}

\item{\code{edges}}{\code{data.frame}. Connections between electrodes represented as edges.
Required columns:
\itemize{
\item \code{from} (\emph{integer}): Starting node (electrode ID).
\item \code{to} (\emph{integer}): Ending node (electrode ID).
\item \code{weight} (\emph{numeric}): Weight of the connection (e.g., distance or strength).
}}

\item{\code{meta}}{\code{list}. Metadata about the experiment, including:
\itemize{
\item \code{duration}: Duration of the experiment (e.g., in seconds).
\item \code{time_interval}: Sampling time interval for the recordings.
}}

\item{\code{images}}{\code{list}. A list of associated images (e.g., snapshots of MEA recordings).}

\item{\code{name}}{character. String that provides an informative label about the condition
under which the measurements were conducted. (e.g. the drug inside the medium).}

\item{\code{nodes}}{\code{data.frame}. Node-level information about the electrodes. Required columns:
\itemize{
\item \code{id}: Electrode ID.
\item \code{x_orig}: X-coordinate before scaling to image resolution.
\item \code{y_orig}: Y-coordinate before scaling to image resolution.
}}

\item{\code{obj_info}}{A list containing object information such as default instructions and directories.}

\item{\code{sample}}{character. Name of the tissue sample from which the MEA data was generated.}

\item{\code{stimulations}}{\code{data.frame}. Stimulation event data. Columns include:
\itemize{
\item \code{id}: Electrode ID where stimulation occurred.
\item \code{instance}: Time point of stimulation.
\item \code{start}: Start time of stimulation.
\item \code{end}: End time of stimulation.
\item \code{amplitude}: Amplitude of the stimulation.
}}

\item{\code{time_series}}{\code{data.frame}. Time-series data for each electrode. Columns include:
\itemize{
\item \code{time}: Time points.
\item \code{id}: Electrode ID.
\item \code{value}: Recorded signal or firing rate.
}}

\item{\code{version}}{\code{list}. A list containing version information.}
}}

\section{Difference between name and sample}{

Slot \verb{@sample} refers to a name for the tissue sample used. Slot \verb{@name} refers
to the specific MEA data set that has been created according to the set up
captured by the \code{\link{MeaDesign}} object in slot \verb{@MeaDesign}. This distinction is necessary,
since one sample (one tissue slide) can undergo multiple MEA experiments with
different set ups, particularly with respect to the condition/drug applied to the
medium as well as to time resolution.
}

