% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{nElectrodes}
\alias{nElectrodes}
\alias{nMeaNodes}
\alias{nMeaNodes,SPATA2-method}
\alias{nMeaNodes,MeaData-method}
\title{Count MEA data content}
\usage{
nElectrodes(object)

nMeaNodes(object, ...)

\S4method{nMeaNodes}{SPATA2}(object, mea_name = activeMea(object))

\S4method{nMeaNodes}{MeaData}(object, mea_name = activeMea(object))
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}, an object that contains MEA data
or, in case of S4 generics, objects of classes for which a methods has been defined.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}
}
\value{
Numeric value.
}
\description{
Returns the number of instances of miscellaneous aspects in the \code{\link{MeaData}}
object.
}
