% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{passNodeDataToSpata}
\alias{passNodeDataToSpata}
\title{Pass node data to SPATA2 observations}
\usage{
passNodeDataToSpata(
  object,
  var_names,
  mea_name = activeMea(object),
  max_dist = Inf,
  naming = "{var_name}",
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{var_names}{A character vector of variable names from the MEA nodes to transfer to
the SPATA2 observations.}

\item{mea_name}{Character value. The name of the \code{\link{MeaData}} object of interest.}

\item{max_dist}{\link[=concept_distance_measure]{Distance measure}. Maximum allowable distance for assigning a
nearest electrode. Observations beyond this distance are marked as "out-of-range". Defaults to half of
the average electrode spacing.}

\item{naming}{Character value. A glue-compatible template for renaming transferred variables.
Defaults to \code{"{var_name}"}. Renaming must not change the number of variable names.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\description{
This function maps metadata from MEA nodes to observations in a \code{\link{SPATA2}} object
based on the nearest electrode assignment. The metadata is added as new variables to the
\code{\link{SPATA2}} object.
}
\details{
The function first identifies the nearest electrode for each observation in the SPATA object using
\code{\link[=identifyNME]{identifyNME()}}. Metadata variables from the MEA nodes, specified in \code{var_names}, are then
transferred to the observations. Observations farther than \code{max_dist} from any electrode are excluded
from the mapping. Transferred variables can be renamed using the \code{naming} argument, allowing for
custom naming patterns in the \code{\link{SPATA2}} metadata.
}
\seealso{
\code{\link[=identifyNME]{identifyNME()}}, \code{\link[=addMetaDataObs]{addMetaDataObs()}}, \code{\link[=getMeaNodes]{getMeaNodes()}}
}
