% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{passSpataMetaToNodes}
\alias{passSpataMetaToNodes}
\title{Pass SPATA2 metadata to MEA nodes}
\usage{
passSpataMetaToNodes(
  object,
  var_names,
  mea_names = getMeaNames(object),
  max_dist = Inf,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{var_names}{\emph{character}. A vector of metadata variable names in the \code{\link{SPATA2}} object to transfer
to the MEA nodes.}

\item{mea_names}{\emph{character} (optional). A vector of MEA names to which the metadata should be
transferred. Defaults to all MEAs in the object (\code{getMeaNames(object)}).}

\item{max_dist}{\emph{numeric} (optional). Maximum allowable distance for mapping metadata from
observations to nodes. Observations beyond this distance are excluded from the mapping.
Default is \code{Inf}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\description{
This function maps metadata variables from a \code{\link{SPATA2}} object to MEA nodes, based on
the nearest electrode (NME) for each observation. The mapped metadata is added as variables
to the nodes in the MEA design.
}
\details{
For each MEA specified in \code{mea_names}, this function identifies the nearest electrode (NME)
for each observation using \code{\link[=identifyNME]{identifyNME()}}. Metadata variables specified in \code{var_names} are
then transferred to the corresponding nodes. If multiple observations map to the same node,
the closest observation (based on distance) is used. Observations beyond \code{max_dist} are
excluded from the mapping. The function supports transferring variables to multiple MEAs
within the same \code{\link{SPATA2}} object.
}
\seealso{
\code{\link[=identifyNME]{identifyNME()}}, \code{\link[=addVarsToNodes]{addVarsToNodes()}}, \code{\link[=computeDistToNME]{computeDistToNME()}}
}
