% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{readMeaNodes}
\alias{readMeaNodes}
\title{Read and set MEA nodes}
\usage{
readMeaNodes(object, dir, var_id, ...)
}
\arguments{
\item{object}{An object of class \code{\link{MeaData}}.}

\item{dir}{\emph{character}. The directory of the file containing the node data.}

\item{var_id}{\emph{character}. The name of the variable in the file that corresponds to the node
or electrode IDs.}

\item{...}{Additional arguments passed to \code{\link[=read_mea_file]{read_mea_file()}}.}
}
\value{
A modified \code{MeaData} object with the updated node data.
}
\description{
This function reads node data from an external file, validates the data, and sets
it as the node information in a \code{MeaData} object. The function checks for overlap with the MEA
design and handles missing data for expected electrodes.
}
\details{
\itemize{
\item The function validates the presence of the specified variable (\code{var_id}) in the file and ensures
that the IDs are unique.
\item Nodes that do not overlap with the expected MEA design electrodes are flagged as missing.
\item The node data is joined with the MEA design's electrode positions to create the final node data frame.
}
}
\examples{
# Read nodes from a file and set them in the MeaData object
mea_data <- readMeaNodes(
  object = mea_data,
  dir = "path/to/nodes_file.csv",
  var_id = "electrode_id"
)

}
\seealso{
\code{\link[=read_mea_file]{read_mea_file()}}, \code{\link[=setMeaNodes]{setMeaNodes()}}, \code{\link[=getMeaDesign]{getMeaDesign()}}
}
