% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{setMeaData}
\alias{setMeaData}
\title{Set MEA data}
\usage{
setMeaData(object, mea_data, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets \code{\link{MeaData}} objects.
}
