% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synaptic.R
\name{subsetByMeaOutline}
\alias{subsetByMeaOutline}
\title{Subset data by MEA outline}
\usage{
subsetByMeaOutline(
  object,
  strictly = TRUE,
  expand_outline = eSpacing(object, unit = "px")/2,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or \code{\link{MeaData}}.}

\item{strictly}{Logical value. If \code{TRUE}, only observations strictly inside the MEA outline
are considered. Default is \code{TRUE}.}

\item{expand_outline}{\link[=concept_distance_measures]{Distance measure}. The distance by which to expand the MEA
outline. Can be specified in pixels or other distance units. Defaults to half the spacing
between electrodes (\code{eSpacing(object, unit = "px") / 2}).}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Subsets observations in a \code{SPATA2} or \code{MeaData} object based on their spatial location
relative to the Microelectrode Array (MEA) outline. Observations falling within or outside the
outline, optionally expanded by a buffer, are retained.
}
\seealso{
\code{\link[=getMeaOutline]{getMeaOutline()}}, \code{\link[=eSpacing]{eSpacing()}}, \code{\link[=buffer_area]{buffer_area()}}
}
